+++ tests/test_app_platform.py
@@ -7,7 +7,8 @@ from pykube import ConfigMap
 from pytest_mock import MockerFixture
 
 import pytest_helm_charts
-import pytest_helm_charts.test_runtime_fixtures
+import pytest_helm_charts.api.fixtures
+import pytest_helm_charts.fixtures
 import pytest_helm_charts.giantswarm_app_platform.utils
 import pytest_helm_charts.giantswarm_app_platform.fixtures
 from pytest_helm_charts.clusters import Cluster
@@ -33,7 +34,7 @@ def test_app_factory_working(kube_cluster: Cluster, app_factory: AppFactoryFunc,
     mocker.patch("pytest_helm_charts.giantswarm_app_platform.utils.AppCR", autospec=True)
     mocker.patch("pytest_helm_charts.giantswarm_app_platform.utils.ConfigMap", autospec=True)
     mocker.patch("pytest_helm_charts.giantswarm_app_platform.app.wait_for_apps_to_run", autospec=True)
-    mocker.patch("pytest_helm_charts.fixtures.ensure_namespace_exists", autospec=True)
+    mocker.patch("pytest_helm_charts.api.fixtures.ensure_namespace_exists", autospec=True)
     test_configured_app: ConfiguredApp = app_factory(
         app_name,
         app_version,
@@ -61,7 +62,7 @@ def test_app_factory_working(kube_cluster: Cluster, app_factory: AppFactoryFunc,
     cast(unittest.mock.Mock, app_cm.create).assert_called_once()
 
     # assert that app was created
-    cast(unittest.mock.Mock, pytest_helm_charts.fixtures.ensure_namespace_exists).assert_called_once_with(
+    cast(unittest.mock.Mock, pytest_helm_charts.api.fixtures.ensure_namespace_exists).assert_called_once_with(
         kube_cluster.kube_client, app_namespace, None, None
     )
     app_cr = cast(unittest.mock.Mock, pytest_helm_charts.giantswarm_app_platform.utils.AppCR)
